Add Vault support
I created a EconomyComponent that extends a QuartzComponent but I'm
unfortunately not fully across the QuartzLib pattern so I will probably
need some guidance there.

The base functionality should be included to withdraw money from
player's accounts if they enable economy and set costs.

What I would like to include next is notifications to the player so they
know how much has been removed or if an error occured. I would also like
to include a confirm/deny option which tells the player how much the
operation is going to cost them before they do it.

I'm uncertain of the best way to go about telling the player how much
a copy of a poster is going to cost before they take it.